requirements.txt
flask
pinecone-client
langchain
openai
tiktoken
python-dotenv